Create km.py
