+++ src/pymafia/datatypes/modifier.py
@@ -4,9 +4,7 @@ from dataclasses import dataclass, field
 from functools import partial
 from typing import Any, ClassVar
 
-from pymafia.kolmafia.kolmafia import on_kolmafia_start
-
-from pymafia.kolmafia import km
+from pymafia.kolmafia import km, on_kolmafia_start
 from pymafia.lazy_enum import LazyEnum
