+++ src/pymafia/datatypes/vykea.py
@@ -4,9 +4,7 @@ from dataclasses import dataclass, field
 from functools import partial
 from typing import TYPE_CHECKING, Any, ClassVar
 
-from pymafia.kolmafia.kolmafia import on_kolmafia_start
-
-from pymafia.kolmafia import km
+from pymafia.kolmafia import km, on_kolmafia_start
 from pymafia.lazy_enum import LazyEnum
 
 if TYPE_CHECKING:
